requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dill==0.3.9
distro==1.9.0
docling==2.28.4
docling-core==2.25.0
docling-ibm-models==3.4.1
docling-parse==4.0.1
dotenv==0.9.9
easyocr==1.7.2
et_xmlfile==2.0.0
filelock==3.18.0
filetype==1.2.0
frozenlist==1.5.0
fsspec==2025.3.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
imageio==2.37.0
importlib_metadata==8.6.1
Jinja2==3.1.6
jiter==0.9.0
jsonlines==3.1.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
latex2mathml==3.77.0
lazy_loader==0.4
litellm==1.65.4.post1
lxml==5.3.2
markdown-it-py==3.0.0
marko==2.1.3
MarkupSafe==3.0.2
mdurl==0.1.2
mpire==2.10.2
mpmath==1.3.0
multidict==6.2.0
multiprocess==0.70.17
networkx==3.4.2
ninja==1.11.1.4
numpy==2.2.4
openai==1.72.0
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pluggy==1.5.0
propcache==0.3.1
pyclipper==1.3.0.post6
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
Pygments==2.19.1
pylatexenc==2.10
pypdfium2==4.30.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rpds-py==0.24.0
rtree==1.4.0
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.15.2
semchunk==2.2.2
shapely==2.1.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sympy==1.13.1
tabulate==0.9.0
tifffile==2025.3.30
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.6.0+cpu
torchvision==0.21.0+cpu
tqdm==4.67.1
transformers==4.51.1
typer==0.12.5
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
XlsxWriter==3.2.2
yarl==1.19.0
zipp==3.21.0
